move context to own file
