+++ packages/lab/src/dialog-next/index.tsx
@@ -2,3 +2,4 @@ export * from "./DialogNext";
 export * from "./DialogNextTitle";
 export * from "./DialogNextActions";
 export * from "./DialogNextContent";
+export * from "./DialogNextContext";
